requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
fear-greed>=0.4.0
google-generativeai>=0.7.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
